copied solution to start removing features, renamed both solutions accordingly
