+++ SpaceInvaders-Complete/SpaceInvaders/Resources/ship.jpg
