--- SpaceInvaders/SpaceInvaders.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26730.15
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SpaceInvaders", "SpaceInvaders\SpaceInvaders.csproj", "{EE2F2852-EC2C-4EB1-9BE4-1D6AA5FE79DE}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{EE2F2852-EC2C-4EB1-9BE4-1D6AA5FE79DE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EE2F2852-EC2C-4EB1-9BE4-1D6AA5FE79DE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EE2F2852-EC2C-4EB1-9BE4-1D6AA5FE79DE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EE2F2852-EC2C-4EB1-9BE4-1D6AA5FE79DE}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {77255A5C-58CB-40F3-BEA6-58A5D5A83D7D}
-	EndGlobalSection
-EndGlobal
